+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [unreleased patch version]
+## [unreleased]
 
 ### Added
 
@@ -16,7 +16,7 @@
 * Fixed removal of PostgreSQL extensions (`timescaledb`, `pgcrypto`) by function `truncate_database` triggered on reindex.
 * Fixed updating relation between index and head in DB.
 * Fixed creation of missing project package.
-* Fixed race condition occurred when receiving WS messages in an unexpected order.
+* Fixed race condition caused by event emitter concurrency.
 
 ## 3.0.1 - 2021-09-24